Refactor: ExtensionString 파일에 split메서드와 관련된 test메서드 이동
