+++ Calculator/CalculatorTests/CalculatorTests.swift
@@ -9,13 +9,13 @@ import XCTest
 @testable import Calculator
 
 final class CalculatorTests: XCTestCase {
-
+    
     var sut: LinkedList!
     override func setUpWithError() throws {
         try super.setUpWithError()
         sut = LinkedList()
     }
-
+    
     override func tearDownWithError() throws {
         try super.tearDownWithError()
         sut = nil
@@ -53,29 +53,5 @@ final class CalculatorTests: XCTestCase {
         
         //than
         XCTAssertNotNil(input.next)
-    }
-    
-    func test_StringExtension의_split에_String이_String배열로_나오는지() {
-        // given
-        var input = "StringTest"
-        
-        // when
-        let expectation = ["StringTest"]
-        
-        // then
-        let result = input.split(with: "a")
-        XCTAssertEqual(result, expectation)
-    }
-    
-    func test_StringExtension의_split에_String이_with에_입력된_파라미터값_기준으로_나눠지는지() {
-        // given
-        var input = "String+Test"
-        
-        // when
-        let expectation = ["String","Test"]
-        
-        // then
-        let result = input.split(with: "+")
-        XCTAssertEqual(result, expectation)
     }
 }
